Online Exam Taking System
This is a project for our course Enterprise Application Development. The
project requirements are provided in PDF format.
To run the code, Import the war file in your Eclipse IDE.
The text files are the sample Import files for the Questions and the
Roster.

For guidance, refer to the User Guide and Design Documents.
This project was completed in a group of 2.